backend/requirements.txt
Flask
Flask-CORS
Flask-JWT-Extended
Flask-SQLAlchemy
Werkzeug
python-dotenv
requests